# MCLOUD-5890: Improve Cloud Docker performance