--- a/src/Test/Integration/_files/cloud_base/docker-compose.exp.yml
+++ b/src/Test/Integration/_files/cloud_base/docker-compose.exp.yml
@@ -11,7 +11,7 @@ services:
     ports:
       - '3306'
     volumes:
-      - 'docker-mnt:/mnt:rw,cached'
+      - 'docker-mnt:/mnt:rw,delegated'
       - 'mymagento-mariadb-conf:/etc/mysql/mariadb.conf.d'
       - 'mymagento-magento-db:/var/lib/mysql'
       - 'docker-entrypoint:/docker-entrypoint-initdb.d'
@@ -52,14 +52,14 @@ services:
     image: 'magento/magento-cloud-docker-php:7.3-fpm-1.1'
     extends: generic
     volumes:
-      - 'magento:/app:ro,cached'
-      - 'magento-vendor:/app/vendor:ro,cached'
-      - 'magento-generated:/app/generated:ro,cached'
-      - 'magento-var:/app/var:rw,cached'
-      - 'magento-app-etc:/app/app/etc:rw,cached'
-      - 'magento-pub-media:/app/pub/media:rw,cached'
-      - 'magento-pub-static:/app/pub/static:rw,cached'
-      - 'docker-mnt:/mnt:rw,cached'
+      - 'magento:/app:ro,delegated'
+      - 'magento-vendor:/app/vendor:ro,delegated'
+      - 'magento-generated:/app/generated:ro,delegated'
+      - 'magento-var:/app/var:rw,delegated'
+      - 'magento-app-etc:/app/app/etc:rw,delegated'
+      - 'magento-pub-media:/app/pub/media:rw,delegated'
+      - 'magento-pub-static:/app/pub/static:rw,delegated'
+      - 'docker-mnt:/mnt:rw,delegated'
     networks:
       magento:
         aliases:
@@ -74,14 +74,14 @@ services:
     ports:
       - '80:80'
     volumes:
-      - 'magento:/app:ro,cached'
-      - 'magento-vendor:/app/vendor:ro,cached'
-      - 'magento-generated:/app/generated:ro,cached'
-      - 'magento-var:/app/var:rw,cached'
-      - 'magento-app-etc:/app/app/etc:rw,cached'
-      - 'magento-pub-media:/app/pub/media:rw,cached'
-      - 'magento-pub-static:/app/pub/static:rw,cached'
-      - 'docker-mnt:/mnt:rw,cached'
+      - 'magento:/app:ro,delegated'
+      - 'magento-vendor:/app/vendor:ro,delegated'
+      - 'magento-generated:/app/generated:ro,delegated'
+      - 'magento-var:/app/var:rw,delegated'
+      - 'magento-app-etc:/app/app/etc:rw,delegated'
+      - 'magento-pub-media:/app/pub/media:rw,delegated'
+      - 'magento-pub-static:/app/pub/static:rw,delegated'
+      - 'docker-mnt:/mnt:rw,delegated'
     environment:
       - VIRTUAL_HOST=magento2.docker
       - VIRTUAL_PORT=80
@@ -129,10 +129,10 @@ services:
     image: 'magento/magento-cloud-docker-php:7.3-cli-1.1'
     extends: generic
     volumes:
-      - 'magento:/app:rw,cached'
-      - 'magento-vendor:/app/vendor:rw,cached'
-      - 'magento-generated:/app/generated:rw,cached'
-      - '~/.composer/cache:/root/.composer/cache:rw,cached'
+      - 'magento:/app:rw,delegated'
+      - 'magento-vendor:/app/vendor:rw,delegated'
+      - 'magento-generated:/app/generated:rw,delegated'
+      - '~/.composer/cache:/root/.composer/cache:rw,delegated'
     networks:
       magento-build:
         aliases:
@@ -149,14 +149,14 @@ services:
     image: 'magento/magento-cloud-docker-php:7.3-cli-1.1'
     extends: generic
     volumes:
-      - 'magento:/app:ro,cached'
-      - 'magento-vendor:/app/vendor:ro,cached'
-      - 'magento-generated:/app/generated:ro,cached'
-      - 'magento-var:/app/var:rw,cached'
-      - 'magento-app-etc:/app/app/etc:rw,cached'
-      - 'magento-pub-media:/app/pub/media:rw,cached'
-      - 'magento-pub-static:/app/pub/static:rw,cached'
-      - 'docker-mnt:/mnt:rw,cached'
+      - 'magento:/app:ro,delegated'
+      - 'magento-vendor:/app/vendor:ro,delegated'
+      - 'magento-generated:/app/generated:ro,delegated'
+      - 'magento-var:/app/var:rw,delegated'
+      - 'magento-app-etc:/app/app/etc:rw,delegated'
+      - 'magento-pub-media:/app/pub/media:rw,delegated'
+      - 'magento-pub-static:/app/pub/static:rw,delegated'
+      - 'docker-mnt:/mnt:rw,delegated'
     networks:
       magento:
         aliases: