… src/Test/Integration/_files/custom_cloud_base/docker-compose.exp.yml