<!---
    Thank you for contributing to Magento.
    To help us process this pull request we recommend that you add the following information:
     - Summary of the pull request,
     - Issue(s) related to the changes made,
     - Manual testing scenarios,
-->

<!--- Please provide a general summary of the Pull Request in the Title above -->

### Description
<!---
    Please provide a description of the changes proposed in the pull request.
    Letting us know what has changed and why it needed changing will help us validate this pull request.
-->

Jira task https://jira.corp.magento.com/browse/MCLOUD-5890
There are a few ways to improve the overall performance of Cloud Docker:

https://docs.docker.com/docker-for-mac/osxfs-caching

The comparison chart: https://www.amazee.io/blog/post/docker-on-mac-performance-docker-machine-vs-docker-for-mac

The general idea is to add `:cached` flag to read-only folders so it performs better in real scenarios.

### Results comparison

|   | Before  |  After |
|---|---|---|
| Transactions  |  6221 hits |  6555 hits |
| Availability  | 100.00 %  |  100.00 % |
|  Elapsed time  |  59.94 secs | 59.42 secs  |
| Data transferred | 69.21 MB | 72.80 MB |
| Transaction rate | 103.79 trans/sec | 110.32 trans/sec |
| Throughput | 1.15 MB/sec | 1.23 MB/sec |
| Concurrency | 24.64 | 24.48 |
| Successful transactions | 6221 | 6555 |
| Failed transactions | 0 | 0 |
| Longest transaction | 2.95 | 3.45 |
| Shortest transaction | 0.00 | 0.00 | 

### Fixed Issues (if relevant)
<!---
    If relevant, please provide a list of fixed issues in the format magento/ece-tools#<issue_number>.
    There could be 1 or more issues linked here and it will help us find some more information about the reasoning behind this change.
-->
1. https://jira.corp.magento.com/browse/MCLOUD-5890

### Manual testing scenarios
<!---
    Please provide a set of unambiguous steps to test the proposed code change.
    Giving us manual testing scenarios will help with the processing and validation process.
-->
1. Build developer config
2. Deploy and run smoke tests
3. Build production config
4. Deploy and run smoke tests

### Contribution checklist
 - [ ] Pull request has a meaningful description of its purpose
 - [ ] All commits are accompanied by meaningful commit messages