@andriyShevtsov 2020-05-06
@shiftedreality I can't connect to bash:
```Error response from daemon: OCI runtime create failed: container_linux.go:349: starting container process caused "process_linux.go:449: container init caused \"rootfs_linux.go:58: mounting \\\"/var/lib/docker/volumes/magento-cloud_docker-mnt/_data\\\" to rootfs \\\"/var/lib/docker/overlay2/2da6e18ed51585b25e1e4bd89438975cacf5b7f0ef1f9eb59e52e3a0f7c44e20/merged\\\" at \\\"/var/lib/docker/overlay2/2da6e18ed51585b25e1e4bd89438975cacf5b7f0ef1f9eb59e52e3a0f7c44e20/merged/mnt\\\" caused \\\"no such file or directory\\\"\"": unknown```